+++ internal/core/entities/session.go
@@ -8,16 +8,16 @@ import (
 
 
 type CustomClaims struct {
-	UserID    int64      `json:"sub"`
+	UserID    uint64      `json:"sub"`
 	TokenType string    `json:"token_type"`
 	jwt.RegisteredClaims
 }
 
 type Session struct {
-	ID           uint       `json:"id" gorm:"primaryKey"`
+	ID           uint64     `json:"id" gorm:"primaryKey"`
 	Device       string     `json:"device" gorm:"column:device;type:varchar(255);not null"`
 	RefreshToken string     `json:"refresh_token" gorm:"column:refresh_token;type:text;not null"`
-	UserID       uint       `json:"user_id" gorm:"column:user_id;not null;index"`
+	UserID       uint64     `json:"user_id" gorm:"column:user_id;not null;index"`
 	IpAddress    string     `json:"ip_address" gorm:"column:ip_address;type:varchar(50)"`
 	ExpiresAt    time.Time  `json:"expires_at" gorm:"column:expires_at;not null"`
 	CreatedAt    time.Time  `json:"created_at" gorm:"autoCreateTime"`
