+++ internal/ports/session_repository.go
@@ -9,6 +9,6 @@ import (
 type SessionRepository interface {
 	// Session
 	CreateSession(ctx context.Context, session *entities.Session) (error)
-	DeleteSession(ctx context.Context, session uint) (error)
+	DeleteSession(ctx context.Context, session uint64) (error)
 	GetValidSessionByRefreshToken(ctx context.Context, hashedToken string) (*entities.Session, error)
 }
